rename main class to vTheme add tests, add/update log functions.

+++ src/test/clear.ts
@@ -1,5 +1,7 @@
-export const clearTheme = () => {
+export const clearTheme = (storage = true) => {
   document.head.innerHTML = '';
   document.body.innerHTML = '';
-  localStorage.clear();
+  if (storage) {
+    localStorage.clear();
+  }
 };
